+++ CHANGELOG.rst
@@ -244,6 +244,8 @@ Added
   `#1051 <https://github.com/openego/eGon-data/issues/1051>`_
 * Heat pump desaggregation to buildings
   `PR #903 <https://github.com/openego/eGon-data/pull/903>`_
+* Add low flex scenario 'eGon2035_lowflex'
+  `#822 <https://github.com/openego/eGon-data/issues/822>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 .. _PR #703: https://github.com/openego/eGon-data/pull/703